refactor: admin/program 폴더 내부 구조 리펙토링 및 import 경로 수정
- live/, challenge/ 서브폴더 생성 및 이동
- SelectFormControl ui/form/으로 이동
