+++ src/domain/admin/pages/ChallengeCreate.tsx
@@ -4,11 +4,11 @@
 import { fileType, uploadFile } from '@/api/file';
 import { usePostChallengeMutation } from '@/api/program';
 import ChallengePreviewButton from '@/domain/admin/ChallengePreviewButton';
-import ChallengeBasic from '@/domain/admin/program/ChallengeBasic';
-import ChallengeCurriculum from '@/domain/admin/program/ChallengeCurriculum';
-import ChallengeOptionSection from '@/domain/admin/program/ChallengeOptionSection';
-import ChallengePoint from '@/domain/admin/program/ChallengePoint';
-import ChallengePrice from '@/domain/admin/program/ChallengePrice';
+import ChallengeBasic from '@/domain/admin/program/challenge/ChallengeBasic';
+import ChallengeCurriculum from '@/domain/admin/program/challenge/ChallengeCurriculum';
+import ChallengePoint from '@/domain/admin/program/challenge/ChallengePoint';
+import ChallengeOptionSection from '@/domain/admin/program/challenge/ChallengeOptionSection';
+import ChallengePrice from '@/domain/admin/program/challenge/ChallengePrice';
 import ProgramBestReview from '@/domain/admin/program/ProgramBestReview';
 import ProgramBlogReviewEditor from '@/domain/admin/program/ProgramBlogReviewEditor';
 import ImageUpload from '@/domain/admin/program/ui/form/ImageUpload';
@@ -35,7 +35,7 @@ import dynamic from 'next/dynamic';
 import { useRouter } from 'next/navigation';
 import { useCallback, useState } from 'react';
 import { FaSave } from 'react-icons/fa';
-import ChallengeLecture from '../program/ChallengeLecture';
+import ChallengeLecture from '../program/challenge/ChallengeLecture';
 import ChallengeFaqCategory from './program/ChallengeFaqCategory';
 import ChallengeMentorRegistrationSection from './program/ChallengeMentorRegistrationSection';
 import ProgramSchedule from './program/ProgramSchedule';
