+++ src/domain/admin/pages/ChallengeEdit.tsx
@@ -15,11 +15,11 @@ import {
 } from '@/api/program';
 import ChallengePreviewButton from '@/domain/admin/ChallengePreviewButton';
 import EditorApp from '@/domain/admin/lexical/EditorApp';
-import ChallengeBasic from '@/domain/admin/program/ChallengeBasic';
-import ChallengeCurriculumEditor from '@/domain/admin/program/ChallengeCurriculum';
-import ChallengeOptionSection from '@/domain/admin/program/ChallengeOptionSection';
-import ChallengePointEditor from '@/domain/admin/program/ChallengePoint';
-import ChallengePrice from '@/domain/admin/program/ChallengePrice';
+import ChallengeBasic from '@/domain/admin/program/challenge/ChallengeBasic';
+import ChallengeCurriculumEditor from '@/domain/admin/program/challenge/ChallengeCurriculum';
+import ChallengePointEditor from '@/domain/admin/program/challenge/ChallengePoint';
+import ChallengeOptionSection from '@/domain/admin/program/challenge/ChallengeOptionSection';
+import ChallengePrice from '@/domain/admin/program/challenge/ChallengePrice';
 import ProgramBestReview from '@/domain/admin/program/ProgramBestReview';
 import ProgramBlogReviewEditor from '@/domain/admin/program/ProgramBlogReviewEditor';
 import ImageUpload from '@/domain/admin/program/ui/form/ImageUpload';
@@ -45,7 +45,7 @@ import { useQueryClient } from '@tanstack/react-query';
 import { useParams, useRouter } from 'next/navigation';
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { FaSave } from 'react-icons/fa';
-import ChallengeLecture from '../program/ChallengeLecture';
+import ChallengeLecture from '../program/challenge/ChallengeLecture';
 import ChallengeFaqCategory from './program/ChallengeFaqCategory';
 import ChallengeMentorRegistrationSection from './program/ChallengeMentorRegistrationSection';
 import ProgramSchedule from './program/ProgramSchedule';
