+++ src/domain/admin/pages/LiveCreate.tsx
@@ -3,10 +3,10 @@
 import { fileType, uploadFile } from '@/api/file';
 import { usePostLiveMutation } from '@/api/program';
 import LivePreviewButton from '@/domain/admin/LivePreviewButton';
-import LiveBasic from '@/domain/admin/program/LiveBasic';
-import LiveCurriculum from '@/domain/admin/program/LiveCurriculum';
-import LiveMentor from '@/domain/admin/program/LiveMentor';
-import LivePrice from '@/domain/admin/program/LivePrice';
+import LiveCurriculum from '@/domain/admin/program/live/LiveCurriculum';
+import LiveMentor from '@/domain/admin/program/live/LiveMentor';
+import LivePrice from '@/domain/admin/program/live/LivePrice';
+import LiveBasic from '@/domain/admin/program/live/LiveBasic';
 import ProgramBestReview from '@/domain/admin/program/ProgramBestReview';
 import ProgramBlogReviewEditor from '@/domain/admin/program/ProgramBlogReviewEditor';
 import ImageUpload from '@/domain/admin/program/ui/form/ImageUpload';
@@ -28,7 +28,7 @@ import { FaSave } from 'react-icons/fa';
 import ProgramSchedule from './program/ProgramSchedule';
 
 const LiveInformation = dynamic(
-  () => import('@/domain/admin/program/LiveInformation'),
+  () => import('@/domain/admin/program/live/LiveInformation'),
   { ssr: false },
 );
