+++ src/domain/admin/pages/program/ChallengeMentorRegistrationSection.tsx
@@ -4,7 +4,7 @@ import {
   useAdminChallengeMentorListQuery,
   useAdminUserMentorListQuery,
 } from '@/api/mentor/mentor';
-import SelectFormControl from '@/domain/admin/program/SelectFormControl';
+import SelectFormControl from '@/domain/admin/program/ui/form/SelectFormControl';
 import Heading2 from '@/domain/admin/ui/heading/Heading2';
 import {
   Checkbox,
