+++ src/domain/admin/program/challenge/ChallengeCurriculum.tsx
