+++ src/domain/admin/program/challenge/ChallengeLecture.tsx
