+++ src/domain/admin/program/challenge/ChallengePoint.tsx
