+++ src/domain/admin/program/guidebook/GuidebookBasicSection.tsx
@@ -1,5 +1,5 @@
 import Input from '@/common/input/v1/Input';
-import SelectFormControl from '@/domain/admin/program/SelectFormControl';
+import SelectFormControl from '@/domain/admin/program/ui/form/SelectFormControl';
 import {
   CreateGuidebookReq,
   liveAndVodJob,
