+++ src/domain/admin/program/live/LiveCurriculum.tsx
